.history/requirements_20251109232603.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
typesense
langchain_openai
langgraph
langchain_text_splitters